Fix type definitions in studio/index.d.ts (#5227)
**User-Facing Changes**
Fixed some invalid TypeScript types in the `@foxglove/studio` package.

**Description**
Use an `index.ts` source file with a build step, rather than
hand-writing `index.d.ts`, to ensure TypeScript actually checks for
errors in CI.
