+++ packages/studio-base/src/players/FoxgloveWebSocketPlayer/index.ts
@@ -42,6 +42,7 @@ import {
   ServiceCallPayload,
   ServiceCallRequest,
   ServiceCallResponse,
+  Parameter,
 } from "@foxglove/ws-protocol";
 
 import WorkerSocketAdapter from "./WorkerSocketAdapter";
@@ -728,7 +729,7 @@ export default class FoxgloveWebSocketPlayer implements Player {
     }
 
     log.debug(`FoxgloveWebSocketPlayer.setParameter(key=${key}, value=${value})`);
-    this._client.setParameters([{ name: key, value }], uuidv4());
+    this._client.setParameters([{ name: key, value: value as Parameter["value"] }], uuidv4());
 
     // Pre-actively update our parameter map, such that a change is detected if our update failed
     this._parameters.set(key, value);
