+++ packages/studio-base/src/players/Ros1Player.ts
@@ -547,8 +547,6 @@ export default class Ros1Player implements Player {
 
   public setParameter(key: string, value: ParameterValue): void {
     log.debug(`Ros1Player.setParameter(key=${key}, value=${value})`);
-    // seems to be a TypeScript issue - the ParameterValue type is treated as `any`
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
     void this._rosNode?.setParameter(key, value);
   }
