+++ packages/studio/tsconfig.json
@@ -1,0 +1,12 @@
+{
+  "extends": "@foxglove/tsconfig/base",
+  "include": ["./**/*"],
+  "compilerOptions": {
+    "composite": true,
+    "incremental": true,
+    "rootDir": "./src",
+    "outDir": "./dist",
+    "noEmit": true,
+    "lib": ["es2020"]
+  }
+}
